entryページを削除し、editを作成
